Update `encoders` with more clear functionality
Through a code audit we've discovered a number of implementations of
DecimalEncoder. Some of these have different semantics. This can lead to
hard-to-understand code or very subtle bugs.

This code adds new functions with names that better match the intent of
the commit. We'll use those in place of all the legacy implementations
within the project.