WIP: Aligning key elements to `common`
The bulk of this commit has been porting over code from `common` into
pycommon, looking for bugs, adding unit tests for coverage, etc. This is
not yet ready to go into dev for two primary reasons:

1. It has not yet been thoroughly tested in local testing
2. The unit tests are not at 100% and will ultimately fail

We're at near 85% at this moment and should be able to get the rest of
the way shortly so long as we don't add any other features from common
yet.

The plan is to get the rest of what is there covered by tests, and do
some actual testing with people, then merge into Dev. From there we can
begin moving other pieces into this.