Remove salt from TokenV1, add tests, add makefile (#3)
* Update requirements.txt

* WIP: Aligning key elements to `common`

The bulk of this commit has been porting over code from `common` into
pycommon, looking for bugs, adding unit tests for coverage, etc. This is
not yet ready to go into dev for two primary reasons:

1. It has not yet been thoroughly tested in local testing
2. The unit tests are not at 100% and will ultimately fail

We're at near 85% at this moment and should be able to get the rest of
the way shortly so long as we don't add any other features from common
yet.

The plan is to get the rest of what is there covered by tests, and do
some actual testing with people, then merge into Dev. From there we can
begin moving other pieces into this.

* Get coverage to 100% on current branch

* Test bugfix & add API Token generation

Added a new api_utils file that has some key generation utilities.
Also fixed a bug with one unit test coming back with an unordered list.

* Add `setup` directive to Makefile

This directive sets up the Python virtual environment, installs the
requirements, sets the scripts to executable, and runs the pre-commit hook
script.

* Remove salt from TokenV1, update tests

After chatting with the team, we determined that we do not need to salt
the API keys because of the inherent 256bit high entry strength. This
change updates TokenV1 to remove the salt.