Fix tools test (#8)
* Fix tests for tools/*

The test_tools.py didn't maintain 100% coverage due to an idiosyncracy
of the ArgumentParser. The lines in `main` that check if the command is
either 'ls' or 'register' (if/elif) have an implicit "neither of these"
condition that our coverage wanted to satisfy.
However, argparse throws and error if our `choices` of `ls` or
`register` weren't provided. Therefore, we added an explicit else block
with no coverage (since its unreachable) and add some documentation.

* Update github action job to align with Makefile

* Revert "Update github action job to align with Makefile"

This reverts commit 6b1ede3291be5de3e7faeff1f4e055d1468cc89e.

* Unify test and coverage branch differences