Introduce DAL Framework (#21)
* initial commit: work in progress DAL branch

This branch (not yet covered) is a work in progress branch to
introduce the DAL interface to pycommon. Lots of work still needs
to be done to tease out the value here.

Complete work in progress.

* Some refactoring work and adding functionality

This commit remains a work in progress and is not to be used for
anything other than ongoing development.

* prevent isort from incorrectly reordering

the dal/providers/aws __init__ is currently critical in its ordering. If
it were to be reordered, things break. Add statements to make isort not
do that.

* Fix tests, add tests, more functionality added

* Minor function optimization

* Minor bugfixes

* Fixup some config items and clean lint failures

* Add unit tests

* fix coverage

* Add accounts delete when user is deleted

When a user object is deleted, we now look for any associated account
objections associated with them and delete those too.

* Fix Account Decimal bug

* Fix Account Decimal bug

* Add DAL README

* Remove hard-coded table value from AwsAccount and AwsUser