Bugfixes, Add UserSettings stub, add Tests
This commit fixes some environment variable issues in the AwsUser
and AwsAccount classes. It also adds a _UserSettings class in the
AwsUser file. This class is not a "Top Level" class (in the same
way that AwsUser) because conceptually, the Settings _belong to_
the user and the DAL doesn't need to formally describe the data-
structures implementing this are defined. More concretely: if a
user decides to implement Amplify in SQLite3, we don't want them to
have to use the exact same data model. What we do want is for the
code to work, so they must keep the semantics the same, though.

I also added tests for the `update_username` functionality of the
_UserSettings class.