Jwks and dal updates (#22)
* Add memoization to the JWKS retrieval process

Since we'll need the JWKS as part of the account addition/upgrade
process to validate our tokens, this change adds a memoization process
to reduce the number of calls. It also (more importantly) splitsout the
JWKS retrieval functions to make them available outside of the
get_claims() function.

There are also some additions to the DAL in this branch which is
not yet in-play. Those are bugfixes, tests, etc.

* Updated DAL tests and bug fixes