Add REMOVE semantics to save() on User
If a User object has one of its attributes set to None, and that
attribute is NOT NONE upstream, then this change will ensure that
the attribute is removed from the record. For example, if a
record has cust_saml_groups = ["ABC"] and we run the following
code:
```python
user = AwsUser(...)
user.cust_saml_groups = None
user.save()
```
The attribute will be removed in the DynamoDB record.